backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-CORS==5.0.0
Flask-Mail==0.9.1
PyMySQL==1.1.1
python-dotenv==0.21.1
PyJWT==2.8.0
Werkzeug==2.2.3
blinker==1.7.0

# Dépendances pour la protection PDF contre téléchargeurs automatiques
pytest==7.4.3
pytest-flask==1.3.0
PyYAML==6.0.1

# Note: PDF.js est chargé depuis CDN, pas besoin d'installation locale
